Add files via upload
The Structure Hijack sample shows some very strange voodoo when using a pointer to a void, cast or not, data still comes out of it. This was done with VS2017 and didn't seem to mind it but  I would not bet C compilers on other systems will be so forgiving.